backend/mtc_etedjmodel/scripts/conexion_qda.py
import encode
import pymysql
import pymysql.cursors

print('hola mundo')

# Connect to the database
connection = pymysql.connect(host='SAM_SQL03',
                             user='qda',
                             password='passwd',
                             db='QDA_DAIMLER',
                             charset='qda',
                             cursorclass=pymysql.cursors.DictCursor)
try:
    with connection.cursor() as cursor:
        # Create a new record
        sql = "INSERT INTO `users` (`email`, `password`) VALUES (%s, %s)"
        cursor.execute(sql, ('webmaster@python.org', 'very-secret'))

    # connection is not autocommit by default. So you must commit to save
    # your changes.
    connection.commit()

    with connection.cursor() as cursor:
        # Read a single record
        sql = "SELECT `id`, `password` FROM `users` WHERE `email`=%s"
        cursor.execute(sql, ('webmaster@python.org',))
        result = cursor.fetchone()
        print(result)
finally:
    connection.close()
