backend/mtc_etedjmodel/requirements.txt
django==2.1.2
djangorestframework==3.10.3
psycopg2==2.7.5
psycopg2-binary==2.8.3
pandas==0.25.1
PyYAML==5.1.2
python-decouple==3.1
dj-database-url==0.5.0
django-url-filter==0.3.13
celery==4.3.0
django_celery_beat==1.5.0
django_celery_results==1.1.2
eventlet==0.25.1
redis==3.3.8
requests==2.22.0
gunicorn==19.9.0
django-extensions==2.2.3
pillow==6.2.0
django-filter==1.1.0
whoosh==2.7.4
django-haystack==2.8.1
Pympler==0.7
numpy==1.17.3
